utils: time formatting function to YANG
Add a function to print time_point as timezone-aware string in format
accepted by yang:date-and-time type. Also add a function to convert the
string representation back to time_point.

For simpler dealing with chrono objects and timezone-aware datetimes
(C++20 library functions dealing with those are still not fully
implemented in libstdc++) we decided to use HowardHinnant/date [1]
library.

[1] https://github.com/HowardHinnant/date

Change-Id: I39480e0424cb862e12df329820f9778fef35e04b