daemon: support alarm's leaf time-created
Alarms have a leaf that states when the alarm entry was created. Quoting
RFC 8632 [1]:

> The timestamp when this alarm entry was created.  This
> represents the first time the alarm appeared; it can
> also represent that the alarm reappeared after a purge.
> Further state changes of the same alarm do not change
> this leaf; these changes will update the 'last-changed'
> leaf.

There was small implementation challenge in the testing phase. The
daemon should publish the current time to sysrepo so we couldn't
hardcode the times in the tests. There were two possibilities of solving
that:
1) Create custom TrivialClock [2] that will be manipulated manually
   from the tests. This would probably require templating Daemon class
   with the TrivialClock used (our custom or the std::system_clock).
2) Let daemon publish current time in sysrepo and deal with that in the
   tests, probably by allowing some time interval to appear in the
   data.

We chose option number 2. Class AnyTimeBetween representing allowed
time_point interval was created and custom comparators comparing with
string YANG values.

[1] https://datatracker.ietf.org/doc/html/rfc8632
[2] https://en.cppreference.com/w/cpp/named_req/TrivialClock

Change-Id: Iad97981cde4632194cbebaed8ca004b5753acd2f