daemon: support alarm's leaf last-changed
Alarms have a leaf that states when the alarm entry was last raised.
Quoting RFC 8632 [1]:

> A timestamp when the 'status-change' or 'operator-state-change' list
> was last changed.

The 'status-change' and 'operator-state-change' lists are both in
different 'if-feature' blocks. However, from description of 'list
status-change' we can deduce what actions change this leaf:
 - changed severity (warning, minor, major, critical),
 - changed clearance status, and
 - alarm-text update.

The operator-state is not relevant to changes of leaves we implement
now.

[1] https://datatracker.ietf.org/doc/html/rfc8632

Change-Id: I1748d9459e3a19e4df2fdd454a70f21731d17a31