daemon: send alarm inventory notifications
The ietf-alarms YANG model defines one should send a notification
whenever alarm-inventory list changes.

Populating the alarm-list is a responsibility of the client and we don't
publish any interface. The client just pushes the data in the
operational datastore. There might be a slight problem: When the
client disconnects, the data in the alarm-inventory list disappear,
hence there might be some alarms without an entry in the
alarm-inventory. We will deal with that later (if necessary).

Change-Id: I42fe9ad87535a669a7c45bdea0f7702a6a0826a5