tests: rework time interval testing
The daemon writes to sysrepo current times. In testing, we have to
handle that somehow (see a1d99f1833d293424cda88d482d6ce0fec3cb2de for
rationale).
However, the tests were testing whether the time is in the time
interval between the RPC call and 300ms after. This time windows is too
much benevolent and it effectively tested only whether the real time
"was not before" the interval start (see previous commit for a bug
caused by this).

This has an easy fix. The RPCs are executed synchronously so let's store
timepoints right before and right after the RPC invocation and expect
that the times that were supposed to be changed by this RPC operation
are in this interval.

Change-Id: I64103a8f7b1e8b9448b7268f0abc0536aa5bad66