daemon: put alarms into shelved-alarms list if appropriate
We start working on alarm-shelving feature of the ietf-alarms model [1].
This implements basic checking if alarm (identified by its keys) belongs
to any configured shelf. Also when a new alarm is created it's put in the
shelf if it belongs there.

This commit however does not implement moving alarms between lists
(shelved and "ordinary" lists) when shelf configuration changes.

[1]: https://datatracker.ietf.org/doc/html/rfc8632

Change-Id: I509a6eadfd37a91c896dd05fa4381ed0c7fb7315