utils: RAII-like sysrepo datastore switcher
In our codebase there seem to be recurring pattern for temporary switch
of a sysrepo session to another datastore which is really verbose and
also not exception-safe.

This commit introduces a scoped datastore switcher that makes provided
sysrepo session switch to a provided datastore and when the object is
destroyed the session will switch back to the original datastore.

Change-Id: I7068cdbcb8f9affa0cf38006eef4feb775e3b929