alarms: implement validate alarms against alarm-inventory
The ietf-alarms works with the idea that the list of alarms is
predetermined. Then, alarm operators can prepare a workflow for managing
all the alarms, because they know what alarms can appear in advance.
According to RFC 8632 the system MUST publish all possible alarm types
in the alarm-inventory container (see RFC 8632 [1, sec. 4.2]).

We have strictly implemented that. Any new alarm that is to be published
is rejected if the corresponding entry is not found in the
alarm-inventory.

[1] https://datatracker.ietf.org/doc/html/rfc8632

Change-Id: I4555f93fd26837f19b0e5020f5c8ac2dbee82d24