Use atomic discards for the operational DS
Port away from sysrepo::Connection::discardOperationalChanges() because
sysrepo now supports discarding via the usual sr_edit_batch(), yay.

I decided to change our method so that it takes an explicit
libyang::Context. I was thinking that this should be extractable from a
libyang::DataNode, but in fact that would have to go through the
associated schema node, and because of the possibility of opaque nodes
and other mess there might not be a schema node at all.

Depends-on: https://gerrit.cesnet.cz/c/CzechLight/dependencies/+/6243
Change-Id: Ic2d9c3769d4cbed0796baa7dcf7a8dbb1cb7ad8f