sysrepo: subscribe to all alarm-control data
The code only created an active subscription (within the running DS) for
the shelving configuration. There are two other leafs in there, each of
them with a default value, which did not have an active subscription
even though they were actually "in effect" -- the code which raised
alarms consulted these two leafs.

Ensure that these leafs are visible in the operational DS by widenging
the scope of the subscription.

Change-Id: I5b4e6b45e9e4afa494680581eb07996a46f6134c