notifications: cache current settings
Now that the subscription is in place, let's avoid one extra read from
sysrepo. Note that the read was already very cheap because it was not
hitting the operational DS (which is big in the benchamrk when there are
"many alarms"). Originally, I did this in two patches because I wanted
to first switch to reading from the data tree that was just read from
the operational DS (so it was just a matter of passing some
libyang::DataNode instances around), but it is easier to read this way.

There's no real speedup here, but it felt like a trivial change.
Basically, there's no point in asking libyang for whatever the current
notification settings are when these could be easily cached in two
simple variables, right?

After implementing that, I spent two hours trying to figure out WTF the
test-alarm_notifications was failing on GCC13, non-sanitizer build. This
is what the failure said:

 /home/jkt/work/cesnet/gerrit/CzechLight/sysrepo-ietf-alarms/tests/alarm_notifications.cpp:131: FATAL ERROR: REQUIRE( lastChangedTimesInSysrepo == lastChangedTimesFromNotifications ) is NOT correct!
   values: REQUIRE( {
   "2024-01-22T15:26:06.547436758+01:00",
   "2024-01-22T15:26:06.552589315+01:00",
   "2024-01-22T15:26:06.559164264+01:00",
   "2024-01-22T15:26:06.567906182+01:00",
   "2024-01-22T15:26:06.574688440+01:00",
 } == {
   "2024-01-22T15:26:06.552589315+01:00",
   "2024-01-22T15:26:06.559164264+01:00",
   "2024-01-22T15:26:06.567906182+01:00",
   "2024-01-22T15:26:06.574688440+01:00",
 } )

 /home/jkt/work/cesnet/gerrit/CzechLight/sysrepo-ietf-alarms/tests/alarm_notifications.cpp:79: ERROR: Pending expectation on destroyed mock object:
 Expected eventsAlarmStatus.notified(trompeloeil::eq(expectProps)) to be called once, actually never called
   param  _1 == {
   "alarm-text": "Hey, I'm overheating.",
   "alarm-type-id": "alarms-test:alarm-1",
   "perceived-severity": "warning",
   "resource": "edfa",
   "time": "(AnyTimeBetween [2024-01-22T15:26:06.547065430+01:00] and [2024-01-22T15:26:06.847065430+01:00])",
 }

Of course the only frakkin' reason was that the subscriber did not have
the `sysrepo::SubscribeOptions::Enabled` flag set, and I was not
initializing that particular member, so it was consulting some random
crap value. You might have told me, ASAN, you know?

Anyway, this has no performance impact, but it will help us in future
when we cease reading the operational data.

Change-Id: I3b6d2fbe12121bfd455b91bab9c531fe94ae165e