refactor: distinguish between "alarm type" and "alarm key"
My interpretation of the "alarm inventory" is that the lists of
resources and severities are optional. I would like to add a cached copy
of the alarm inventory, and the most obvious approach to that is
treating the pair of (type ID, type qualifier) as a key which indexes
the "allowed resources" and "allowed severities". Let's split the `Key`
into these two parts.

I'm not sure whether to make the `alarm-type-qualifier` optional or not.
The current code simply shuffles empty strings around, but the YANG
model sometimes allows that value to simply not be present. This will
have to wait for a follow-up commit, in the meanwhile, just adding a
FIXME.

[1] https://datatracker.ietf.org/doc/html/rfc8632#section-4.2

Change-Id: I1fa0b836d8eeffe28ed967ff925340562124c7db