refactor: distinguish between "alarm type" and "alarm key"
My interpretation of the "alarm inventory" is that the lists of
resources and severities are optional. I would like to add a cached copy
of the alarm inventory, and the most obvious approach to that is
treating the pair of (type ID, type qualifier) as a key which indexes
the "allowed resources" and "allowed severities". Let's split the `Key`
into these two parts.

I was a bit confused whether the `alarm-type-qualifier` is an optional
string or not. The YANG model actually uses the word "optional", but
considering that this leaf is actually a key of a list, it must be there
as per YANG 1.1 rules. Of course this is all about the operational DS
where the concept of "validity" is rather relaxed, but hey, let's stick
with empty strings now.

[1] https://datatracker.ietf.org/doc/html/rfc8632#section-4.2

Change-Id: I1fa0b836d8eeffe28ed967ff925340562124c7db