Move alarm inventory to C++
Don't fetch the inventory info from sysrepo anytime an alarm is
created/updated. Instead, remember inventory updates and use that cached
information at the time an alarm is raised.

This is actually pretty inefficient; it slows the current benchmark down
by about 6.7%. The reason is that anytime an inventory entry is pushed,
the entire operational DS of this module has to be parsed. That sucks,
but it will enable much faster operation once we could get away with
that DS read at alarm update time. Also, stray tuned for the next patch
which will switch this to lazy loading.

Change-Id: I633d7bc5484e2a7fd8b555f640a6dbddb73d895f