lazy population of alarm inventory cache
The parent commit introduced an annoying performance hog into the code
path where alarm inventory entries were processed. That was done in the
name of a possible future speedup of "regular" alarm handling. The
bottleneck was actually in that dreaded read from sysrepo; the code
would have to load and parse the module's operational DS content each
time the inventory changed.

Considering that the original issue was that "pushing inventory entries
is slow", making it even slower is not good at all. Solve that by *not*
populating the inventory cache until it's really needed. This speeds up
the code by about 6%, which makes it almost on par with how it was when
no inventory cache was maintained.

Change-Id: I9ce5a5bce9025950e4a0257272baad85f7015b56