statistics: maintain cache of all alarms
This commit introduces a local cache which holds "all alarms", no matter
whether they are supposed to be shelved or not, and whether they have
been already cleared or if they are still active. This cache is then
used for creating statistics without re-fetching the data from sysrepo.
This avoids a read/modify/write cycle of the entire operational DS for
this module, and compared to the previous version, there's an
incremental speedup of 43%.

The data are still being read each time an alarm is updated. This will
be removed in some future commit when the alarm handling code will be
only pushing stuff to sysrepo, without reading the data all the time.

The change in the test suite looks valid to me. On line 85, the test
clears an alarm that's currently shelved. Because such as change is
visible in the `shelved-alarms/shelved-alarm` list, I believe that doing
that should touch the `shelved-alarms-last-changed` leaf. Previously,
the test actually checked that the timestamp is set to *before* that
last RPC (which clears that shelved alarm) was triggered.

Change-Id: Ie40a852d1605f55a41e307059737251fb26cf9cc