benchmark: only measure the final purge
...because we are not really interested in the overhead of the sysrepo
library when a client calls an action. Also, that time is spent in the
*client* which is requesting that RPC/action (which is good), and it's
something that's completely out-of-control from us. There surely is
still a bottleneck, the sysrepo code reads the operational DS content
for the `ietf-alarms` module, and that's going to be slow, but hey, we
have to live with that.

Numbers-wise, this "only" introduces about a 2% overhead compared to the
status when no purging actions were triggered at all.

Change-Id: Ic6fc0bd03486c8cdd5b34e3d06fed60454d80856