avoid AlarmEntry implicit construction in the cache
std::map::operator[](const Key&) default-constructs or zero-initializes
when the key is not present [1].
This is what we use for getting alarms from m_alarms cache. However,
when the alarm was cleared, then the function returned in the next few
lines and left the alarm zero-initialized in the m_alarms map.

This had an unfortunate consequence when somebody called the alarm RPC
with an cleared alarm (this had to be the first call with such key) and
then updateStatistics was triggered, there was an alarm in the cache
with lastSeverity=0 because of the zero-initialization.

And that was the full story behind this error seen in our velia CI logs.

  libyang[0]: Invalid enumeration value "_". (path: Schema location "/ietf-alarms:alarms/summary/alarm-summary/severity".)

[1] https://en.cppreference.com/w/cpp/container/map/operator_at

Fixes: cbddff3dcad568fdcf8b5d5d70e4696c8b4295dd
Change-Id: I82fee8c300244902081c93d3be510a5fbd92ba60