tests: a more useful logging pattern
Rather than showing just "second.fractions", which might, if we're very
unlucky, hide the fact that a test has just slept for a minute or two,
use a built-in formatter that shows the relative difference to the
previous message. I'm keeping the resolution in milliseconds.

And since we have a custom pattern, enable color formatting. Note that
by default, it (still) won't color format when wrapped in a `ctest` run,
which means that the CI log won't contain "terminal color" escapes. I
guess that's what we want (at least in the CI), but not necessarily in
the terminal.

Change-Id: Ie7de02eea3d07b82a7b5d8dff8ec5e7644f5ee7f