shelving: cache the control data
Using the already-cached data about shelving rules means that there's
one fewer access to sysrepo when raising alarms. That access was already
pretty cheap because it's against the `running` DS, not `operational`,
but hey, why not.

There's no significant speedup, but it "cleans up" the hot path of alarm
raising.

Change-Id: I283065db68297af08c13b0320a90150c704ff189