tests: fix relative time pattern in the logs
Guess what this log snippet means:

 +214ms [2433 main I] Will notify about any alarm state changes
 +6147ms [2433 main E] Sequentially pushing 200 inventory entries: 6050ms
 +191ms [2433 main I] Will notify about any alarm state changes
 +231ms [2433 main E] Pushing 200 inventory entries at once: 131ms
 +41277ms [2433 main E] Sending 200 alarms: 41277ms
 +44043ms [2433 sysrepo E] EV ORIGIN: SHM event "rpc" ID 1 processing timed out.
 +0ms [2433 sysrepo W] EV ORIGIN: "/ietf-alarms:alarms/alarm-list/purge-alarms" "rpc" ID 1 priority 0 failed (Timeout expired).
 +1ms [2433 sysrepo E] User callback failed.

I thought that it (obviously) says that `EV ORIGIN: SHM event "rpc" ID 1
processing timed out` happened roughly 44s after the `Sending 200
alarms: 41277ms`, but that's not the case. Calling `spdlog::set_pattern`
actually propagates the pattern to all the "loggers" (sources or
producers of log messages). All of these individual loggers, such as
`main` and `sysrepo` in the example above, were then configured to use
relative timestamps. The problem was that these timestamps were actually
relative to the "previous message within the same logger", and not to
the message shown on the previous line. Funny, eh?

Fix that by setting the relative timestamps at the *sink* level.

Fixes: 2244686 tests: a more useful logging pattern
Change-Id: I829bca5773f86347daaee66907d62401cc6b223f