do not invalidate the alarm inventory cache
The cache was being rebuilt over and over again, even when there were no
changes to the inventory since the last time. What a silly bug.

The synthetic benchmark is 25% faster compared to the previous version.
Compared to the very beginning, it's 66% faster (and we're measuring an
extra purge which was not included when this benchmarking started).

Fixes: 2d15c1c lazy population of alarm inventory cache
Change-Id: I0ba443eafdcb36e32794c014fef57e5f2795b6b8