performance: trace how much time is spent in sysrepo interactions
I did not want to mess with passing a custom logger, so we now have a
default one (the "main" one, obviously). I'm very much tempted to get
rid of all the custom logging, TBH, but that's for some other time.

The explicit flush is there to avoid lost performance data on (some)
exceptions.

This is taken from velia. It introduces a new dependency on `fmt` (which
is already used by `spdlog`, so it isn't entirely a "new dependency").

See-Also: I7d0a4d79e4777f77544ad0f39af8ce58de2c31a6
Change-Id: If522bff34cc8952778c76b78ffb1fdd3b0964248