tests: remove terminating semicolon from a macro definition
We use CLIENT_ALARM_RPC macro in a way that it raises an alarm and it
returns a lower and upper bound times when the alarm was actually
raised.

The definition of the macro ends with a semicolon. That causes problem
when somebody wants to use the macro in a code like this:

  vec.emplace_back(CLIENT_ALARM_RPC(args...));

... and of course, we want to use code like this in the following
commits.

Change-Id: I1fe4f0c21148d71ddebf3023eab02d5a02a53a19