implement alarm's status-change list
This is the initial commit that starts the implemention of alarm-history
feature from ietf-alarms model [1].

We start with simple addition of alarm changes to the status-change list
that is contained in every alarm entry. The length of the history list is
controlled by the control/max-alarm-status-changes leaf entry.

This commit does not yet implement actions for compressing the alarm
history.

[1] https://datatracker.ietf.org/doc/rfc8632/

Change-Id: I5b89db4fe8c3a12dcb34c46415c303288de949d4