sysrepo: new approach to pushed ops data
There's been a change upstream, each session now has its own edit which
is used to build the final content of the operational DS. As a result,
we have to stop using the old approach with "discard" nodes. Instead,
there's now a "global" edit which is always re-applied upon any change.

Change-Id: Iff7c473933e5e75c618077230c9bb81337c19269
Depends-on: https://gerrit.cesnet.cz/c/CzechLight/dependencies/+/8134