sysrepo: new approach to pushed ops data
There's been a change upstream, each session now has its own edit which
is used to build the final content of the operational DS. As a result,
we have to stop using the old approach with "discard" nodes. Instead,
there's now a "global" edit which is always re-applied upon any change.

While we're at this, let's simplify the code a bit; don't log too much
information, and performs some small refactorings, align the data types
for stats reporting with the YANG model, etc.

I tried to do this in a series of commits, but to be fair, it did not
read that nicely, and I got fed up with the endless rebases.

Change-Id: Iff7c473933e5e75c618077230c9bb81337c19269
Depends-on: https://gerrit.cesnet.cz/c/CzechLight/dependencies/+/8134